\name{qlcTokenize-package}
\alias{qlcTokenize-package}
\alias{qlcTokenize}
\docType{package}
\title{
Orthography processing using Orthography Profiles
}
\description{
This packages uses a general approach to describe orthographic systems through so-called Orthography Profiles. It offers functions to write such profiles based on some actual written text, and to test and correct profiles given concrete data. The main end-use is to produce tokenized texts in so-called tailored grapheme clusters.
}
\details{
\tabular{ll}{
Package: \tab qlcTokenize\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-10-28\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Michael Cysouw <cysouw@mac.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
