\name{qlcTokenize-package}
\alias{qlcTokenize-package}
\alias{qlcTokenize}
\docType{package}
\title{
Orthography processing using Orthography Profiles
}
\description{
This packages uses a general approach to describe orthographic systems through so-called Orthography Profiles. It offers functions to write such profiles based on some actual written text, and to test and correct profiles given concrete data. The main end-use is to produce tokenized texts in so-called tailored grapheme clusters.
}
\details{
\tabular{ll}{
Package: \tab qlcTokenize\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2015-04-03\cr
License: \tab GPL-3\cr
}
The central function is \code{\link{tokenize}}. A basic sceleton for an orthography profile can be produced with \code{\link{write.profile}}
}
\author{
Michael Cysouw <cysouw@mac.com>
}
\keyword{ package }

